move the logic to CodeSanityValidator
